pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigseg"
version = "1.0.0"
description = "vigseg: Vision Transformer analysis pipeline for kidney tissue samples."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Christos Botos", email = "botoschristos@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "numpy>=2.2",
    "scipy>=1.14",
    "pandas>=2.2",
    "matplotlib>=3.10",
    "pillow>=10.0",
    "imageio>=2.33",
    "scikit-image>=0.23",
    "scikit-learn>=1.7",
    "joblib>=1.5",
    "tqdm>=4.67",
    "torch>=2.7",
    "torchvision>=0.22",
    "anndata>=0.11",
    "scanpy>=1.11",
    "transformers>=4.52",
    "psutil>=7.0",
    "pyarrow>=20.0",
    "fastparquet>=2024.2",
    "imagecodecs>=2025.3",
    "pywavelets>=1.8",
    "contourpy>=1.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
gpu = ["cupy-cuda12x>=13.5"]

[project.scripts]
vigseg = "vigseg.cli:main"

[project.urls]
Repository = "https://github.com/ChrisBotos/vigseg"

[tool.setuptools.packages.find]
where = ["src"]
include = ["vigseg*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
